requirements.txt
biopython
pandas
numpy
scikit-learn
joblib
matplotlib
seaborn
tqdm
